Fix: correct TimeKeeper Algorithm
TimeKeeper algorithm was wrong. For a given participant, it should
compare the lastCalculated time, with the max time of the current
timeHistorList.  A given participant may have a persistently bad
time. They should never be allowed to go backward, which means
that under certain conditions the timeHistoryList should be
discarded.

Signed-off-by: Kevin O'Donnell <kevin@blockchaintp.com>